UWEBKIT2: Handle 64 bit GPU surface id

+++ uWebKitNative/UWKProcess/uwk_gpusurface.h
@@ -27,7 +27,7 @@ public:
     virtual void UpdateTexture(const void* image_buffer, uint32_t x, uint32_t y, uint32_t width, uint32_t height,
                                uint32_t rowLength, uint32_t skipPixels, uint32_t skipRows) = 0;
 
-    virtual uint32_t GetSurfaceID() = 0;
+    virtual uintptr_t GetSurfaceID() = 0;
 
     static GpuSurface* Create(int maxWidth, int maxHeight);
